requirements.txt
anyascii==0.3.0
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.9.3
black==21.12b0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
decorator==5.1.1
Django==3.2.11
django-filter==21.1
django-google-maps==0.12.4
django-modelcluster==5.2
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
executing==0.8.2
html5lib==1.1
idna==3.3
ipython==8.0.0
jedi==0.18.1
jovian==0.2.41
l18n==2021.3
lxml==4.7.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.22.2
openpyxl==3.0.9
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.1
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.1.4
tablib==3.1.0
telepath==0.2
tomli==1.2.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
uuid==1.30
wagtail==2.15.1
wagtail-blocks==0.7.2
wagtail-maps==0.2.0
wagtailfontawesome==1.2.1
wagtailgeowidget==6.1.0
wagtailgmaps==1.0.1
wagtailleafletwidget==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
